Day_2/requirements.txt
pandas
numpy
scikit-learn
fastapi[standard]
uvicorn
pymongo
python-dotenv
jupyter
joblib
pydantic
requests